Added Admin page, revised BookRepositoryService, revised templates, moved SQLAlchemy models into dedicated Models file.
The admin page can be accessed only by users with isAdmin equal to True. From the Admin page, editing and deleting Users and Books is allowed. Note: currently all users are admins upon account creation for convenience while we develop; if you need to test a regular user account, you can edit an account's admin status from the admin page.

Revised the BookRepositoryService file. Removed the methods for editing and deleting database entries as those functions are now covered by the admin page. Checking in/out and renewing books now works; when books are added to the database, they are displayed on the homepage with options to select these actions under them.
If we want to add the separate models for transactions and fees, they can be added to Models.py; for the time being I only have a one to many relationship set up between the User and Book models.