Added search bar, error check, and AccountingService (empty) directory and file; revamped the way flashes are handled; added some CSS styling
Search bar on the home page filters books by title or author.

Added error check for user registration to check if the username already exists.

Updated the way errors are sent to the pages from string variables passed in with render_template, to flashed messages of category "error". This performs the same task but makes the HTML for displaying them a bit cleaner.